+++ javascript/game/Camera/CameraContract.js
@@ -1,4 +1,21 @@
 class CameraContract {
+    /* -------------------------------------------------------------------------- */
+    /*                                   Methods                                  */
+    /* -------------------------------------------------------------------------- */
+    GetControlledCamera() { }
+
+    GetMatrixWorldInverse() { }
+
+    GetProjectionMatrix() { }
+
+    GetPosition() { }
+
+    GetQuaternion() { }
+
+    GetWorldPosition() { }
+
+    GetWorldDirection() { }
+
     SetAspectRatio(aspectRatio) { }
 }