+++ javascript/user-interface/RootCanvas.jsx
@@ -1,10 +1,11 @@
+import { CelestialMarkerScreen } from "./hover-markers/CelestialMarkerScreen.jsx";
 import { Component } from "react";
-import { SolarSystemVisualizer } from "../SolarSystemVisualizer.js";
-import { LoadingScreen } from "./loading-screen/LoadingScreen.jsx";
 import { EventMediator } from "./mediator/EventMediator.js";
+import { LoadingScreen } from "./loading-screen/LoadingScreen.jsx";
 import { OptionsMenuGroup } from "./option-menu/OptionsMenuGroup.jsx";
-import { ViewingOptionsGroup } from "./option-menu/ViewingOptionsGroup.jsx";
+import { SolarSystemVisualizer } from "../SolarSystemVisualizer.js";
 import { TimeControlScreen } from "./time-control/TimeControlScreen.jsx";
+import { ViewingOptionsGroup } from "./option-menu/ViewingOptionsGroup.jsx";
 
 class RootCanvas extends Component {
     constructor(props) {
@@ -38,11 +39,14 @@ class RootCanvas extends Component {
         const canvasClassName = `canvas ${this.state.isMenuVisible ? "fullscreen" : ""}`;
 
         return (
-            <div id="root-canvas" className={canvasClassName}>
-                <LoadingScreen />
-                <ViewingOptionsGroup />
-                <OptionsMenuGroup />
-                <TimeControlScreen />
+            <div id="fullScreenCanvas" className="fullscreen">
+                <div id="root-canvas" className={canvasClassName}>
+                    <LoadingScreen />
+                    <ViewingOptionsGroup />
+                    <OptionsMenuGroup />
+                    <TimeControlScreen />
+                </div>
+                <CelestialMarkerScreen />
             </div>
         );
     }