+++ javascript/user-interface/option-menu/options/ShowMarkerButton.jsx
@@ -1,8 +1,23 @@
 import { BaseOptionButton } from "./BaseOptionButton.jsx";
 
 class ShowMarkerButton extends BaseOptionButton {
+    constructor(props) {
+        super(props);
+
+        this.state = {
+            isActive: true,
+            isVisible: false,
+            hasFadedOut: false
+        };
+    }
+
+    /* -------------------------------------------------------------------------- */
+    /*                               Event Handlers                               */
+    /* -------------------------------------------------------------------------- */
+
     HandleClick() {
-        console.log("Show marker behaviour not implemented.");
+        this.eventMediator.Notify("ToggleMarkers");
+        this.setState({ isActive: !this.state.isActive });
     }
 }