+++ javascript/utils/math-library.js
@@ -1,3 +1,4 @@
+import * as THREE from "three";
 import { ObjectValidator } from "./ObjectValidator.js";
 
 class MathHelper {
@@ -28,8 +29,29 @@ class MathHelper {
     static IsNotZero(value) {
         return ObjectValidator.IsValid(value) && value !== 0;
     }
+
+    static WorldSpaceToScreenSpace(worldPosition, canvasDimension, camera) {
+        const screenPosition = new THREE.Vector3();
+        const worldPosition4D = new THREE.Vector4(worldPosition.x, worldPosition.y, worldPosition.z, 1);
+        const clipSpacePosition = worldPosition4D.applyMatrix4(camera.GetMatrixWorldInverse()).applyMatrix4(camera.GetProjectionMatrix());
+        const ndc = new THREE.Vector3(clipSpacePosition.x / clipSpacePosition.w, clipSpacePosition.y / clipSpacePosition.w, clipSpacePosition.z / clipSpacePosition.w);
+
+        screenPosition.x = (ndc.x + 1) / 2 * canvasDimension.width;
+        screenPosition.y = (1 - ndc.y) / 2 * canvasDimension.height;
+
+        return { x: screenPosition.x, y: screenPosition.y };
+    }
+
+    /// <summary>
+    /// Returns the direction from source to target. This uses THREE.js Vector3.
+    /// </summary>
+    static GetDirectionToTarget(source, target) {
+        return new THREE.Vector3().subVectors(target, source).normalize();
+    }
+
+    static GetDistanceBetweenObjects(source, target) {
+        return source.position.distanceTo(target.position);
+    }
 }
 
-export {
-    MathHelper
-};
+export { MathHelper };